+++ hexa/user_management/middlewares.py
@@ -5,8 +5,9 @@ from typing import Callable
 from django.conf import settings
 from django.http.request import HttpRequest
 from django.http.response import HttpResponse
-from django.shortcuts import redirect, render
+from django.shortcuts import redirect
 from django.urls import resolve, reverse
+from django.utils import translation
 from django_otp.middleware import OTPMiddleware
 
 from hexa.app import get_hexa_app_configs
@@ -75,31 +76,29 @@ def login_required_middleware(
     return middleware
 
 
-def accepted_tos_required_middleware(
-    get_response: Callable[[HttpRequest], HttpResponse]
-) -> Callable[[HttpRequest], HttpResponse]:
-    """
-    TOS is mandatory for all:
-        - authenticated user
-        - except routes logout, index and ready endpoint.
-    """
-
-    def middleware(request: HttpRequest) -> HttpResponse:
-        if (
-            request.user.is_authenticated
-            and is_protected_routes(request)
-            and not getattr(request.user, "accepted_tos", False)
-            and settings.USER_MUST_ACCEPT_TOS
-        ):
-            return render(request, "user_management/terms_of_service.html")
-        else:
-            return get_response(request)
-
-    return middleware
-
-
 class TwoFactorMiddleware(OTPMiddleware):
     """
     This must be installed after
     :class:`~django.contrib.auth.middleware.AuthenticationMiddleware`.
     """
+
+
+class UserLanguageMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        user = request.user
+        language = None
+        if user.is_authenticated:
+            language = user.language
+        else:
+            language = translation.get_language_from_request(request)
+
+        translation.activate(language)
+        request.LANGUAGE_CODE = translation.get_language()
+
+        response = self.get_response(request)
+
+        translation.deactivate()
+        return response