+++ hexa/user_management/models.py
@@ -81,7 +81,7 @@ class User(AbstractUser, UserInterface):
     username = None
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     email = EmailField(_("email address"), db_collation="case_insensitive", unique=True)
-    accepted_tos = models.BooleanField(default=False)
+    language = models.CharField(max_length=10, default="en")
 
     objects = UserManager()