+++ hexa/user_management/tests/test_schema.py
@@ -2,6 +2,7 @@ import re
 from unittest.mock import patch
 
 from django.core import mail
+from django.test import override_settings
 from django.utils.http import urlsafe_base64_encode
 from django_otp import user_has_device
 from django_otp.models import Device
@@ -15,6 +16,11 @@ from hexa.user_management.models import (
     MembershipRole,
     Team,
     User,
+)
+from hexa.workspaces.models import (
+    Workspace,
+    WorkspaceInvitation,
+    WorkspaceMembershipRole,
 )
 
 from ..utils import default_device, devices_for_user
@@ -1012,8 +1018,7 @@ class TwoFactorTest(GraphQLTestCase):
     @classmethod
     def setUp(cls):
         cls.USER_REGULAR = User.objects.create_user(
-            "john@bluesquarehub.com",
-            "regular",
+            "john@bluesquarehub.com", "regular", first_name="John"
         )
         cls.USER_WITH_DEVICE = User.objects.create_user(
             "device@bluesquare.com", "device"
@@ -1248,4 +1253,186 @@ class TwoFactorTest(GraphQLTestCase):
 
         self.assertEqual(
             r["data"]["disableTwoFactor"], {"success": True, "errors": None}
+        )
+
+    def test_update_user(self):
+        self.client.force_login(self.USER_REGULAR)
+
+        self.assertEqual(self.USER_REGULAR.language, "en")
+        self.assertEqual(self.USER_REGULAR.first_name, "John")
+        r = self.run_query(
+            """
+                mutation updateUser($input: UpdateUserInput!) {
+                    updateUser(input: $input) {
+                        success
+                        errors
+                        user {
+                            id
+                            email
+                            firstName
+                            language
+                        }
+                    }
+                }
+            """,
+            {"input": {"firstName": "New first name", "language": "fr"}},
+        )
+        self.assertEqual(
+            {
+                "success": True,
+                "errors": [],
+                "user": {
+                    "id": str(self.USER_REGULAR.id),
+                    "email": self.USER_REGULAR.email,
+                    "firstName": "New first name",
+                    "language": "fr",
+                },
+            },
+            r["data"]["updateUser"],
+        )
+        self.USER_REGULAR.refresh_from_db()
+        self.assertEqual(self.USER_REGULAR.first_name, "New first name")
+        self.assertEqual(self.USER_REGULAR.language, "fr")
+
+    @override_settings(LANGUAGES=(("en", "English"), ("fr", "French")))
+    def test_update_user_invalid_language(self):
+        self.client.force_login(self.USER_REGULAR)
+
+        r = self.run_query(
+            """
+                mutation updateUser($input: UpdateUserInput!) {
+                    updateUser(input: $input) {
+                        success
+                        errors
+                        user {
+                            id
+                            language
+                        }
+                    }
+                }
+            """,
+            {"input": {"language": "nl"}},
+        )
+        self.assertEqual(
+            {
+                "success": False,
+                "errors": ["INVALID_LANGUAGE"],
+                "user": None,
+            },
+            r["data"]["updateUser"],
+        )
+
+
+class RegisterTest(GraphQLTestCase):
+    @classmethod
+    def setUp(cls):
+        cls.USER_REGULAR = User.objects.create_user(
+            "john@bluesquarehub.com",
+            "regular",
+        )
+        cls.FEATURE = Feature.objects.create(code="workspaces")
+        cls.WORKSPACE = Workspace.objects.create(name="Workspace")
+        cls.WORKSPACE_INVITATION = WorkspaceInvitation.objects.create(
+            workspace=cls.WORKSPACE,
+            email="johndoe@email.com",
+            role=WorkspaceMembershipRole.EDITOR,
+        )
+
+    def test_register_invalid_token(self):
+        r = self.run_query(
+            """
+            mutation register($input: RegisterInput!) {
+                register(input: $input) {
+                    success
+                    errors
+                }
+            }
+            """,
+            {
+                "input": {
+                    "password1": "Pa$$Word",
+                    "password2": "Pa$$Word",
+                    "firstName": "John",
+                    "lastName": "Doe",
+                    "invitationToken": "zdzd",
+                }
+            },
+        )
+
+        self.assertEqual(
+            r["data"]["register"], {"success": False, "errors": ["INVALID_TOKEN"]}
+        )
+
+    def test_register_password_mismatch(self):
+        r = self.run_query(
+            """
+            mutation register($input: RegisterInput!) {
+                register(input: $input) {
+                    success
+                    errors
+                }
+            }
+            """,
+            {
+                "input": {
+                    "password1": "Pa$$Word",
+                    "password2": "Pa$$Word2",
+                    "firstName": "John",
+                    "lastName": "Doe",
+                    "invitationToken": self.WORKSPACE_INVITATION.generate_invitation_token(),
+                }
+            },
+        )
+
+        self.assertEqual(
+            r["data"]["register"], {"success": False, "errors": ["PASSWORD_MISMATCH"]}
+        )
+
+    def test_register_ok(self):
+        r = self.run_query(
+            """
+            mutation register($input: RegisterInput!) {
+                register(input: $input) {
+                    success
+                    errors
+                }
+            }
+            """,
+            {
+                "input": {
+                    "password1": "Pa$$Word1",
+                    "password2": "Pa$$Word1",
+                    "firstName": "John",
+                    "lastName": "Doe",
+                    "invitationToken": self.WORKSPACE_INVITATION.generate_invitation_token(),
+                }
+            },
+        )
+
+        self.assertTrue(r["data"]["register"]["success"])
+
+        # Check if authenticated
+        r = self.run_query(
+            """
+            query {
+                me {
+                    user {
+                        id
+                        firstName
+                        lastName
+                        email
+                    }
+                }
+            }
+            """,
+        )
+
+        self.assertEqual(
+            r["data"]["me"]["user"],
+            {
+                "id": str(User.objects.get(email=self.WORKSPACE_INVITATION.email).id),
+                "firstName": "John",
+                "lastName": "Doe",
+                "email": self.WORKSPACE_INVITATION.email,
+            },
         )